Move `LegacyDrawer` to `:feature:navigation:drawer` module
